fix bug in syscall.S

+++ include/page.h
@@ -44,7 +44,7 @@ typedef unsigned long pte_t;
 #define PTECNT_PER_PAGE (PAGE_SIZE/sizeof(pte_t))
 
 #define PAGE_ITEMS (PAGE_SIZE/sizeof(unsigned long))
-#define PAGE_ALIGN(page)    (page & PAGE_MASK)
+#define PAGE_ALIGN(page)    (((unsigned long)(page)) & PAGE_MASK)
 #define PAGE_UP(page)     (((unsigned long)page + PAGE_SIZE -1) & PAGE_MASK)
 #define PAGE_DOWN    PAGE_ALIGN
